IN-163 fix posts entity types
