+++ src/entities/comments/api/comments.api.ts
@@ -1,27 +1,24 @@
 import { inctagramApi } from '@/common/api/inct.api'
-import { Likes, Params } from '@/entities/posts'
+import { WithSortPaginationParams } from '@/common/types/api.types'
+import { CommentLikesRequest, CommentLikesResponse } from '@/entities/posts/types/likes.types'
 
-import {
-  AnswerLikesRequest,
-  CommentAnswers,
-  CommentAnswersRequest,
-  CommentLikesRequest,
-  Comments,
-} from '../types/comments.type'
+import { AnswersRequest, AnswersResponse, CommentsResponse } from '../types/comments.types'
 
 export const commentsApi = inctagramApi.injectEndpoints({
   endpoints: builder => ({
-    getAnswerLikes: builder.query<Likes, AnswerLikesRequest>({
-      query: data => {
-        const { answerId, commentId, postId, ...params } = data
+    getAnswerLikes: builder.query<CommentLikesResponse, { answerId: number } & CommentLikesRequest>(
+      {
+        query: data => {
+          const { answerId, commentId, postId, ...params } = data
 
-        return {
-          params,
-          url: `v1/posts/${postId}/comments/${commentId}/answers/${answerId}/likes`,
-        }
-      },
-    }),
-    getCommentAnswers: builder.query<CommentAnswers, CommentAnswersRequest>({
+          return {
+            params,
+            url: `v1/posts/${postId}/comments/${commentId}/answers/${answerId}/likes`,
+          }
+        },
+      }
+    ),
+    getAnswers: builder.query<AnswersResponse, AnswersRequest>({
       query: data => {
         const { commentId, postId, ...params } = data
 
@@ -31,7 +28,7 @@ export const commentsApi = inctagramApi.injectEndpoints({
         }
       },
     }),
-    getCommentLikes: builder.query<Likes, CommentLikesRequest>({
+    getCommentLikes: builder.query<CommentLikesResponse, CommentLikesRequest>({
       query: data => {
         const { commentId, postId, ...params } = data
 
@@ -41,7 +38,7 @@ export const commentsApi = inctagramApi.injectEndpoints({
         }
       },
     }),
-    getComments: builder.query<Comments, { postId: number } & Params>({
+    getComments: builder.query<CommentsResponse, { postId: number } & WithSortPaginationParams>({
       query: data => {
         const { postId, ...params } = data
