+++ src/entities/comments/index.ts
@@ -1,1 +1,2 @@
 export { commentsApi } from './api/comments.api'
+export type { AnswersResponse, CommentsResponse } from './types/comments.types'
