+++ src/entities/posts/index.ts
@@ -1,18 +1,8 @@
 export { postsApi } from './api/posts.api'
 
 export type {
-  Avatar,
-  CreatePostBody,
-  Image,
-  Like,
-  Likes,
-  Params,
-  Post,
-  PostLikes,
-  Posts,
-  PublicPostsRequest,
-  PublicPostsResponse,
-  Reaction,
-  Response,
-  SearchParams,
-} from './types/posts.type'
+  CommentLikesRequest,
+  CommentLikesResponse,
+  PostLikesResponse,
+} from './types/likes.types'
+export type { Post, PublicPostsResponse } from './types/posts.types'
