--- src/entities/posts/types/post.type.ts
@@ -1,28 +1,0 @@
-export type Post = {
-  avatarOwner: string
-  createdAt: string
-  description: string
-  id: number
-  images: Image[]
-  isLiked: boolean
-  likesCount: number
-  location: null | string
-  owner: Owner
-  ownerId: number
-  updatedAt: string
-  userName: string
-}
-
-type Image = {
-  createdAt: string
-  fileSize: number
-  height: number
-  uploadId: string
-  url: string
-  width: number
-}
-
-export type Owner = {
-  firstName: string
-  lastName: string
-}
